+++ boot.py
@@ -176,6 +176,7 @@ cfg = {
         "boot_btn:neg_logic" : True, # boot button gpio0 becomes signaled with value 0 by pressing
         "boot_btn:fire_on_up" : True, # default, fires when releasing
         "boot_btn:event" : "status:mem_1", # event to fire
+        #"boot_btn:event" : "pin:21:toggle", # toggle led on pin 21
 
         "alarm_counter" : None, # not configured
         "alarm_counter:delta_period" : 5,