+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - added module `moddev.control`. listens to events below:
   - `restart` values `[hard|modcore]`. hardreset, or restart all modcore modules (modc)
   - `status` values `[mem_0|mem_1|memfree]`. print the memory info
-  - `gc` . perform gc, print before, and after memory info
+  - `gc` perform gc, print before, and after memory info
+  - `pin`  control output pin, value data format `pinno:[on|off|toggle]`
 - added `EventEmitter` as base class for `Interval`
 - added `AlarmClock` for defined alarm handling (once a day), with optional raising an event
 - ntp module fires `ntp` now also after timezone change